requirements.txt
Cerberus==1.3.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dreamtools-dreamgeeker==1.1.2
Flask==1.1.2
ghostscript==0.6
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.1.0
pyaml==20.4.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
urllib3==1.26.2
Werkzeug==1.0.1

setuptools~=51.1.1
waitress~=1.4.4